requirements-windows.txt
# Flask and Core Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7

# Database - Windows compatible
SQLAlchemy==2.0.21
psycopg2-binary

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# File Storage
boto3==1.28.57
supabase==1.0.4

# AI/ML APIs
google-generativeai==0.3.2
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0

# Email Notifications
sendgrid==6.10.0

# Date/Time Handling
python-dateutil==2.8.2

# Logging and Monitoring
structlog==23.1.0

# Development Dependencies
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Production Dependencies
gunicorn==21.2.0
